// src/api/models/CostsBuiltObject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CostsBuiltObject
 */
export interface CostsBuiltObject {
    /**
     * 
     * @type {string}
     * @memberof CostsBuiltObject
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof CostsBuiltObject
     */
    count: number;
    /**
     * 
     * @type {number}
     * @memberof CostsBuiltObject
     */
    totalSpent: number;
}

/**
 * Check if a given object implements the CostsBuiltObject interface.
 */
export function instanceOfCostsBuiltObject(value: object): value is CostsBuiltObject {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    if (!('totalSpent' in value) || value['totalSpent'] === undefined) return false;
    return true;
}

export function CostsBuiltObjectFromJSON(json: any): CostsBuiltObject {
    return CostsBuiltObjectFromJSONTyped(json, false);
}

export function CostsBuiltObjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): CostsBuiltObject {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'count': json['count'],
        'totalSpent': json['totalSpent'],
    };
}

export function CostsBuiltObjectToJSON(json: any): CostsBuiltObject {
    return CostsBuiltObjectToJSONTyped(json, false);
}

export function CostsBuiltObjectToJSONTyped(value?: CostsBuiltObject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'count': value['count'],
        'totalSpent': value['totalSpent'],
    };
}